Enable xml tests
