+++ tests/jballerina-unit-test/src/test/resources/test-src/types/xml/xml-literals.bal
@@ -6,7 +6,7 @@ function testXMLTextLiteral() returns [xml, xml, xml, xml, xml, xml] {
     xml x2 = xml `${v1}`;
     xml x3 = xml `aaa${v1}bbb${v2}ccc`;
     xml x4 = xml `aaa${v1}bbb${v2}ccc{d{}e}{f{`;
-    xml x5 = xml `aaa${v1}b\${bb${v2}c\}cc{d{}e}{f{`;
+    xml x5 = xml `aaa${v1}b${"\${"}bb${v2}c\}cc{d{}e}{f{`;
     xml x6 = xml ` `;
     return [x1, x2, x3, x4, x5, x6];
 }
@@ -19,7 +19,7 @@ function testXMLCommentLiteral() returns [xml, xml, xml, xml, xml, xml] {
     xml x2 = xml `<!--${v1}-->`;
     xml x3 = xml `<!--aaa${v1}bbb${v2}ccc-->`;
     xml x4 = xml `<!--<aaa${v1}bbb${v2}cccd->e->-f<<{>>>-->`;
-    xml x5 = xml `<!---a-aa${v1}b\${bb${v2}c\}cc{d{}e}{f{-->`;
+    xml x5 = xml `<!---a-aa${v1}b${"\${"}bb${v2}c\}cc{d{}e}{f{-->`;
     xml x6 = xml `<!---->`;
     xml x7 = xml `<!-- $ -->`;
     return [x1, x2, x3, x4, x5, x6];
@@ -34,7 +34,7 @@ function testXMLPILiteral() returns [xml, xml, xml, xml, xml] {
     xml x2 = xml `<?foo ${v1}?>`;
     xml x3 = xml `<?foo  aaa${v1}bbb${v2}ccc?>`;
     xml x4 = xml `<?foo  <aaa${v1}bbb${v2}ccc??d?e>?f<<{>>>?>`;
-    xml x5 = xml `<?foo  ?a?aa${v1}b\${bb${v2}c\}cc{d{}e}{f{?>`;
+    xml x5 = xml `<?foo  ?a?aa${v1}b${"\${"}bb${v2}c\}cc{d{}e}{f{?>`;
     
     return [x1, x2, x3, x4, x5];
 }
@@ -46,7 +46,7 @@ function testExpressionAsAttributeValue() returns [xml, xml, xml, xml, xml] {
     xml x1 = xml `<foo bar="${v0}"/>`;
     xml x2 = xml `<foo bar="aaa${v1}bb'b${v2}ccc?"/>`;
     xml x3 = xml `<foo bar="}aaa${v1}bbb${v2}ccc{d{}e}{f{"/>`;
-    xml x4 = xml `<foo bar1='aaa{${v1}}b\${b"b${v2}c\}cc{d{}e}{f{' bar2='aaa{${v1}}b\${b"b${v2}c\}cc{d{}e}{f{'/>`;
+    xml x4 = xml `<foo bar1='aaa{${v1}}b${"\${"}b"b${v2}c\}cc{d{}e}{f{' bar2='aaa{${v1}}b${"\${"}b"b${v2}c\}cc{d{}e}{f{'/>`;
     xml x5 = xml `<foo bar=""/>`;
     return [x1, x2, x3, x4, x5];
 }
