+++ tests/jballerina-unit-test/src/test/resources/testng-new-parser.xml
@@ -48,16 +48,16 @@
 <!--            <package name="org.ballerinalang.test.structs.*"/>-->
             <package name="org.ballerinalang.test.testerina.*"/>
 <!--            <package name="org.ballerinalang.test.jvm.*"/>-->
-           <package name="org.ballerinalang.test.types.integer"/>
-           <package name="org.ballerinalang.test.types.floattype"/>
-           <package name="org.ballerinalang.test.main.function.*"/>
+            <package name="org.ballerinalang.test.types.integer"/>
+            <package name="org.ballerinalang.test.types.floattype"/>
+            <package name="org.ballerinalang.test.main.function.*"/>
 <!--            <package name="org.ballerinalang.test.types.decimaltype.*"/>-->
             <package name="org.ballerinalang.test.types.any.*"/>
 <!--            <package name="org.ballerinalang.test.error.*"/>-->
 <!--            <package name="org.ballerinalang.test.types.finaltypes.*"/>-->
 <!--            <package name="org.ballerinalang.test.types.finite.*"/>-->
-           <package name="org.ballerinalang.test.types.map.*"/>
-           <package name="org.ballerinalang.test.types.string.*"/>
+            <package name="org.ballerinalang.test.types.map.*"/>
+            <package name="org.ballerinalang.test.types.string.*"/>
 <!--            <package name="org.ballerinalang.test.types.stream.*"/>-->
 <!--            <package name="org.ballerinalang.test.types.table.*"/>-->
 <!--            <package name="org.ballerinalang.test.types.errors.*"/>-->
@@ -65,13 +65,13 @@
 <!--            <package name="org.ballerinalang.test.types.uniontypes.*"/>-->
 <!--            <package name="org.ballerinalang.test.types.typedesc.*"/>-->
 <!--            <package name="org.ballerinalang.test.types.tuples.*"/>-->
-           <package name="org.ballerinalang.test.types.anydata.*"/>
+            <package name="org.ballerinalang.test.types.anydata.*"/>
 <!--            <package name="org.ballerinalang.test.statements.arrays.*"/>-->
 <!--            <package name="org.ballerinalang.test.statements.assign.*"/>-->
             <package name="org.ballerinalang.test.statements.block.*"/>
             <package name="org.ballerinalang.test.statements.breakstatement.*"/>
 <!--            <package name="org.ballerinalang.test.statements.comment.*"/>-->
-<!--            <package name="org.ballerinalang.test.statements.compoundassignment.*"/>-->
+            <package name="org.ballerinalang.test.statements.compoundassignment.*"/>
             <package name="org.ballerinalang.test.statements.continuestatement.*"/>
 <!--            <package name="org.ballerinalang.test.statements.expressionstmt.*"/>-->
 <!--            <package name="org.ballerinalang.test.statements.ifelse.*"/>-->
@@ -92,11 +92,11 @@
 <!--            <package name="org.ballerinalang.test.expressions.conversion.*"/>-->
             <package name="org.ballerinalang.test.expressions.unaryoperations.*"/>
 <!--            <package name="org.ballerinalang.test.expressions.access.*"/>-->
-           <package name="org.ballerinalang.test.expressions.binaryoperations.*"/>
-           <package name="org.ballerinalang.test.expressions.checkedexpr.*"/>
+            <package name="org.ballerinalang.test.expressions.binaryoperations.*"/>
+            <package name="org.ballerinalang.test.expressions.checkedexpr.*"/>
 <!--            <package name="org.ballerinalang.test.expressions.checkpanicexpr.*"/>-->
-           <package name="org.ballerinalang.test.expressions.elvis.*"/>
-           <package name="org.ballerinalang.test.expressions.ternary.*"/>
+            <package name="org.ballerinalang.test.expressions.elvis.*"/>
+            <package name="org.ballerinalang.test.expressions.ternary.*"/>
             <package name="org.ballerinalang.test.expressions.typeof.*"/>
 <!--            <package name="org.ballerinalang.test.expressions.stamp.*"/>-->
 <!--            <package name="org.ballerinalang.test.expressions.lambda.*"/>-->
@@ -110,7 +110,7 @@
 <!--            <package name="org.ballerinalang.test.endpoint.*" />-->
 <!--            &lt;!&ndash;&lt;!&ndash; <package name="org.ballerinalang.test.loc.*" /> &ndash;&gt;&ndash;&gt;-->
             <package name="org.ballerinalang.test.packaging.*" />
-<!--            <package name="org.ballerinalang.test.types.xml.*" />-->
+            <package name="org.ballerinalang.test.types.xml.*" />
 <!--            <package name="org.ballerinalang.test.record.*"/>-->
 <!--            <package name="org.ballerinalang.test.types.constant.*" />-->
 <!--            <package name="org.ballerinalang.test.types.bytetype.*"/>-->
